Add OrangeTypes
