+++ src/orange/CMakeLists.txt
@@ -20,6 +20,7 @@ list(APPEND SOURCES
   OrangeParams.cc
   OrangeParamsOutput.cc
   OrangeTypes.cc
+  OrangeTypesIO.json.cc
   SafetyImager.cc
   detail/BIHBuilder.cc
   detail/BIHPartitioner.cc
