+++ src/orange/OrangeInputIO.json.hh
@@ -8,7 +8,6 @@
 
 #include <nlohmann/json.hpp>
 
-#include "OrangeData.hh"
 #include "OrangeInput.hh"
 
 namespace celeritas
@@ -26,11 +25,6 @@ void to_json(nlohmann::json& j, UnitInput const& value);
 
 void from_json(nlohmann::json const& j, RectArrayInput& value);
 void to_json(nlohmann::json& j, RectArrayInput const& value);
-
-template<class T>
-void from_json(nlohmann::json const& j, Tolerance<T>& value);
-template<class T>
-void to_json(nlohmann::json& j, Tolerance<T> const& value);
 
 void from_json(nlohmann::json const& j, OrangeInput& value);
 void to_json(nlohmann::json& j, OrangeInput const& value);
